requirements.txt
Django==4.0.1
django-cors-headers==3.11.0
django-ninja==0.16.1
dnspython==2.1.0
mongoengine==0.23.1
pymongo==3.12.3


